Create a seperate filter options component
