+++ src/pages/Products.tsx
@@ -2,7 +2,7 @@ import ProductsControl from "../components/ProductsControl"
 
 const Products = () => {
   return (
-    <div className="w-[90%] mx-auto">
+    <div className="w-[90%] mx-auto ">
       <ProductsControl />
     </div>
   )
